Splitting tensornetwork.py file into node.py, edge.py and network.py (#15)
