+++ tensornetwork/edge.py
@@ -1,0 +1,166 @@
+# Copyright 2019 The TensorNetwork Authors
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#      http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+"""Implementation of TensorNetwork structure."""
+
+from __future__ import absolute_import
+from __future__ import division
+from __future__ import print_function
+import collections
+from typing import List, Optional, Union, Text, Tuple
+import numpy as np
+import tensorflow as tf
+import weakref
+from tensornetwork import decompositions
+import Node
+
+
+class Edge:
+  """Edge for the TensorNetwork graph.
+
+  Each edge represents a vector space common to the tensors it connects and over
+  which a contraction may be performed. In numpy terms, each edge represents a
+  `tensordot` operation over the given axes.
+  There are 3 main types of edges:
+
+  Standard Edge:
+    A standard edge is like any other edge you would find in a normal
+    undirected graph as they connect two different nodes. This edge represents
+    a tensor contraction of the underlying tensors along their given axes.
+    The two axes must be the same dimension.
+
+  Dangling Edge:
+    A dangling edge is an edge that only connects to a single node and only one
+    part of the edge connects to the node. The other end is left "dangling".
+    These types of edges can not be contrated and represent additional
+    dimensions on the underlying tensor. After all other edges are contracted,
+    the final result will have the same rank as the number of dangling edges. If
+    there are no dangling edges, then the final value will be a scalar.
+
+  Trace Edges:
+    Trace edges are edges that connects a node to itself. These edges represent
+    a trace along the given axis. Once again, the axes must be the same
+    dimension.
+  """
+
+  def __init__(self,
+               name: Text,
+               node1: Node,
+               axis1: int,
+               node2: Optional[Node] = None,
+               axis2: Optional[int] = None) -> None:
+    """Create an Edge.
+
+    Args:
+      name: Name of the edge. Used primarily for debugging.
+      node1: One of the nodes edge connects.
+      axis1: The axis of node1 that represents this edge.
+      node2: The other node that this edge connects. Can be `None` if edge is
+        dangling.
+      axis2: The axis of node2 that represents this edge. Must be `None` if
+        node2 is `None`.
+
+    Raises:
+      ValueError: If node2 and axis2 are not either both `None` or both
+        not be `None`.
+      tf.errors.InvalidArgumentError: If the dimensions of axis1 and axis2 are
+        not equal.
+    """
+    if (node2 is None) != (axis2 is None):
+      raise ValueError(
+          "node2 and axis2 must either be both None or both not be None")
+    self.name = name
+    self.node1 = node1
+    self.axis1 = axis1
+    self.node2 = node2
+    self.axis2 = axis2
+    self._is_dangling = node2 is None
+
+  def get_nodes(self) -> List[Optional[Node]]:
+    """Get the nodes of the edge."""
+    return [self.node1, self.node2]
+
+  def update_axis(self, old_axis: int, old_node: Node, new_axis: int,
+                  new_node: Node) -> None:
+    """Update the node that Edge is connected to.
+
+    Args:
+      old_axis: The old axis that the edge pointed to.
+      old_node: The old node that the edge pointed to.
+      new_axis: The new axis that the edge should point to.
+      new_node: The new node that replaces the old_node.
+
+    Raises:
+      AssertionError: Whether the edge actually contained `old_node`.
+    """
+    if self.axis1 == old_axis and self.node1 is old_node:
+      self.axis1 = new_axis
+      self.node1 = new_node
+    elif self.axis2 == old_axis and self.node2 is old_node:
+      self.axis2 = new_axis
+      self.node2 = new_node
+    else:
+      raise ValueError("Edge '{}' did not contain node '{}' on axis {}. "
+                       "node1: '{}', axis1: {}, node2: '{}', axis2: {}".format(
+                           self, old_node, old_axis, self.node1, self.axis1,
+                           self.node2, self.axis2))
+  @property
+  def node1(self) -> Node:
+    val = self._node1()
+    if val is None:
+      raise ValueError("node1 for edge '{}' no longer exists.".format(self))
+    return val
+
+  @property
+  def node2(self) -> Optional[Node]:
+    if self._is_dangling:
+      return None
+    if self._node2() is None:
+      raise ValueError("node2 for edge '{}' no longer exists.".format(self))
+    return self._node2()
+  
+  @node1.setter
+  def node1(self, node: Node) -> None:
+    self._node1 = weakref.ref(node)
+
+  @node2.setter
+  def node2(self, node: Optional[Node]) -> None:
+    self._node2 = weakref.ref(node) if node else None
+    if node is None:
+      self._is_dangling = True
+
+  def is_dangling(self) -> bool:
+    """Whether this edge is a dangling edge."""
+    return self._is_dangling
+
+  def is_being_used(self):
+    """Whether the nodes this edge points to also use this edge.
+
+    During edge flattening, nodes can change their edges. Since
+    deleting objects in python isn't possible, we use this to ensure that the
+    edge is actually being used by the given nodes.
+
+    Returns:
+      bool: Whether this edge is actually being used.
+    """
+    result = self is self.node1[self.axis1]
+    if self.node2 is not None:
+      result = result and self is self.node2[self.axis2]
+    return result
+
+  def set_name(self, name):
+    self.name = name
+
+  def __str__(self) -> Text:
+    return self.name
