+++ tensornetwork/node.py
@@ -1,0 +1,235 @@
+# Copyright 2019 The TensorNetwork Authors
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#      http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+"""Implementation of TensorNetwork node structure."""
+
+from __future__ import absolute_import
+from __future__ import division
+from __future__ import print_function
+import collections
+from typing import List, Optional, Union, Text, Tuple
+import numpy as np
+import tensorflow as tf
+import weakref
+from tensornetwork import decompositions
+import Edge
+
+
+class Node:
+  """Node for the TensorNetwork graph.
+
+  A Node represents a concrete tensor in a tensor network. The number of edges
+  for a node represents the rank of that tensor.
+
+  For example:
+  O : No edges means this node represents a scalar value.
+  -O : A single edge means this node is a vector.
+  -O-: Two edges represents a matrix.
+  -O- : Three edges are tensor of rank 3, etc.
+   |
+
+  Each node can have an arbitrary rank/number of edges, each of which can have
+  an arbitrary dimension.
+  """
+
+  def __init__(self,
+               tensor: Union[np.ndarray, tf.Tensor],
+               name: Text,
+               axis_names: List[Text]) -> None:
+    """Create a node for the TensorNetwork.
+
+    Args:
+      tensor: The concrete tensor that is represented by this node. Can be
+        either a numpy array or a tensorflow tensor.
+      name: Name of the node. Used primarily for debugging.
+      axis_names: List of names for each of the tensor's axes.
+
+    Raises:
+      ValueError: If there is a repeated name in `axis_names` or if the length
+        doesn't match the shape of the tensor.
+    """
+    self.tensor = tf.convert_to_tensor(tensor)
+    self.name = name
+    self.edges = [Edge(edge_name, self, i)
+                  for i, edge_name in enumerate(axis_names)]
+    if axis_names is not None:
+      self.add_axis_names(axis_names)
+    else:
+      self.axis_names = None
+
+  def add_axis_names(self, axis_names: List[Text]) -> None:
+    """Add axis names to a Node.
+
+    Args:
+      axis_names: List of names for each of the tensor's axes.
+
+    Raises:
+      ValueError: If there is a repeated name in `axis_names` or if the length
+        doesn't match the shape of the tensor.
+    """
+    if len(axis_names) != len(set(axis_names)):
+      raise ValueError("Not all axis names are unique.")
+    if len(axis_names) != len(self.tensor.shape):
+      raise ValueError("axis_names is not the same length as the tensor shape."
+                       "axis_names length: {}, tensor.shape length: {}".format(
+                           len(axis_names), len(self.tensor.shape)))
+    self.axis_names = axis_names[:]
+
+  def add_edge(self,
+               edge: "Edge",
+               axis: Union[int, Text],
+               override: bool = False) -> None:
+    """Add an edge to the node on the given axis.
+
+    Args:
+      edge: The edge to add.
+      axis: The axis the edge points to.
+      override: If true, replace the existing edge with the new one.
+
+    Raises:
+      ValueError: If the edge on axis is not dangling.
+    """
+    axis_num = self.get_axis_number(axis)
+    if axis_num < 0 or axis_num >= len(self.tensor.shape):
+      raise ValueError("Axis must be positive and less than rank of the tensor")
+    if not self.edges[axis_num].is_dangling() and not override:
+      raise ValueError(
+          "Node '{}' already has a non-dangling edge for axis {}".format(
+              self, axis))
+    self.edges[axis_num] = edge
+
+  def get_tensor(self):
+    return self.tensor
+
+  def set_tensor(self, tensor):
+    self.tensor = tensor
+
+  def reorder_edges(self, edge_order: List["Edge"]) -> "Node":
+    """Reorder the edges for this given Node.
+
+    This will reorder the node's edges and transpose the underlying tensor
+    accordingly.
+
+    Args:
+      edge_order: List of edges. The order in the list determins the new edge
+        ordering.
+
+    Returns:
+      self: This node post reordering.
+
+    Raises:
+      ValueError: If either the list of edges is not the same as expected or if
+        you try to reorder with a trace edge.
+    """
+    if set(edge_order) != set(self.edges):
+      raise ValueError("Given edge order does not match expected edges. "
+                       "Found: {}, Expected: {}".format(edge_order, self.edges))
+    for edge in edge_order:
+      if edge.node1 == edge.node2:
+        raise ValueError("Edge reordering does not support trace edges. "
+                         "Found trace edge: '{}'".format(edge))
+
+    permutation = []
+    for i, edge in enumerate(edge_order):
+      # This is O(n^2), but the number of edges will likely never be >100
+      # so this should be fine for now.
+      old_position = self.edges.index(edge)
+      permutation.append(old_position)
+      edge.update_axis(old_position, self, i, self)
+    self.edges = edge_order[:]
+    self.tensor = tf.transpose(self.tensor, perm=permutation)
+    if self.axis_names is not None:
+      # Update axis_names:
+      tmp_axis_names = []
+      for i in permutation:
+        tmp_axis_names.append(self.axis_names[i])
+      self.axis_names = tmp_axis_names
+    return self
+
+  def reorder_axes(self, perm: List[int]) -> "Node":
+    """Reorder axes of the node's tensor.
+
+    This will also update all of the node's edges.
+
+    Args:
+      perm: Permutation of the dimensions of the node's tensor.
+
+    Returns:
+      self: This node post reordering.
+    """
+    if set(perm) != set(range(len(self.edges))):
+      raise ValueError("A full permutation was not passed. "
+                       "Permutation passed: {}".format(perm))
+    self.tensor = tf.transpose(self.tensor, perm=perm)
+    tmp_edges = []
+    for i, position in enumerate(perm):
+      edge = self.edges[position]
+      edge.update_axis(position, self, i, self)
+      tmp_edges.append(edge)
+    self.edges = tmp_edges
+    return self
+
+  def get_axis_number(self, axis: Union[Text, int]) -> int:
+    """Get the axis number for a given axis name or value."""
+    if isinstance(axis, int):
+      return axis
+    try:
+      return self.axis_names.index(axis)
+    except ValueError:
+      raise ValueError("Axis name '{}' not found for node '{}'".format(
+          axis, self))
+
+  def get_dimension(self, axis: Union[Text, int]) -> int:
+    """Get the dimension on the given axis.
+
+    Args:
+      axis: The axis of the underlying tensor.
+
+    Returns:
+      dimension: The dimension of the given axis.
+
+    Raises:
+      ValueError: if axis isn't an int or if axis is too large or small.
+    """
+    axis_num = self.get_axis_number(axis)
+    if axis_num < 0 or axis_num >= len(self.tensor.shape):
+      raise ValueError("Axis must be positive and less than rank of the tensor")
+    return tf.shape(self.tensor)[axis_num]
+
+  def get_edge(self, axis: Union[int, Text]) -> "Edge":
+    axis_num = self.get_axis_number(axis)
+    return self.edges[axis_num]
+
+  def get_all_edges(self):
+    # Copy to prevent overwriting.
+    return self.edges[:]
+
+  def get_all_nondangling(self):
+    """Return the set of nondangling edges connected to this node."""
+    return {edge for edge in self.edges if not edge.is_dangling()}
+
+  def set_name(self, name):
+    self.name = name
+
+  def has_nondangling_edge(self):
+    for e in self.edges:
+      if not e.is_dangling():
+        return True
+    return False
+
+  def __getitem__(self, key: Union[int, Text]) -> "Edge":
+    return self.get_edge(key)
+
+  def __str__(self) -> Text:
+    return self.name
